mariana/pyproject.toml
[project]
name = "mariana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.4",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
]
